pyproject.toml
[tool.poetry]
name = "flaskapicodguns"
version = "0.1.0"
description = ""
authors = ["Owen owenek123@gmail.com"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.0.1"
bs4 = "^0.0.1"
flask-restful = "^0.3.8"
flask-caching = "^1.9.0"
flask-pymongo = "^2.3.0"
flask-mongoengine = "^1.0.0"
pymongo = "^3.11.3"
isodate = "^0.6.0"
werkzeug = "^2.0.1"
dnspython = "^2.1.0"
marshmallow = "^3.11.1"
webargs = "^8.0.0"
flask-cors = "^3.0.10"
passlib = "^1.7.4"
flask-jwt-extended = "^4.1.0"
toml = "^0.10.2"
python-dotenv = "^0.18.0"
opencv-python = "^4.5.5"
flask-smorest = "^0.37.0"
flask-marshmallow = "^0.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
